outsource matrix/vector machinery to Arrays module - rename UnitVector -> SingleEntryVector - rename delete_zero_columns -> delete_zero_columns! - create Arrays module - move matrix and vector machinery to Arrays module
