+++ src/Approximations/decompositions.jl
@@ -554,8 +554,8 @@ julia> B = BallInf(zeros(4), 1.0);
 
 julia> c = constraints_list(B)[1:2]
 2-element Array{HalfSpace{Float64,VN} where VN<:AbstractArray{Float64,1},1}:
- HalfSpace{Float64,LazySets.Approximations.UnitVector{Float64}}([1.0, 0.0, 0.0, 0.0], 1.0)
- HalfSpace{Float64,LazySets.Approximations.UnitVector{Float64}}([0.0, 1.0, 0.0, 0.0], 1.0)
+ HalfSpace{Float64,LazySets.Arrays.SingleEntryVector{Float64}}([1.0, 0.0, 0.0, 0.0], 1.0)
+ HalfSpace{Float64,LazySets.Arrays.SingleEntryVector{Float64}}([0.0, 1.0, 0.0, 0.0], 1.0)
 
 julia> P = HPolyhedron(c);
