+++ src/Approximations/template_directions.jl
@@ -35,49 +35,6 @@ end
 # ==================================================
 
 """
-    UnitVector{T} <: AbstractVector{T}
-
-A lazy unit vector with arbitrary one-element.
-
-### Fields
-
-- `i` -- index of non-zero entry
-- `n` -- vector length
-- `v` -- non-zero entry
-"""
-struct UnitVector{T} <: AbstractVector{T}
-    i::Int
-    n::Int
-    v::T
-end
-
-function Base.getindex(e::UnitVector{T}, i::Int) where T
-    @boundscheck @assert 1 <= i <= e.n
-    return i == e.i ? e.v : zero(T)
-end
-
-Base.size(e::UnitVector) = (e.n,)
-
-function Base.:(*)(A::AbstractMatrix{N}, e::UnitVector{N}) where {N}
-    return A[:, e.i] * e.v
-end
-
-function Base.:(*)(A::Transpose{N, <:AbstractMatrix{N}}, e::UnitVector{N}
-                  ) where {N}
-    return A[:, e.i] * e.v
-end
-
-# diagonal matrix times unit vector
-function Base.:(*)(D::Diagonal{N, V}, e::UnitVector{N}) where {N, V<:AbstractVector{N}}
-    return UnitVector(e.i, e.n, D.diag[e.i] * e.v)
-end
-
-function inner(e1::UnitVector{N}, A::AbstractMatrix{N}, e2::UnitVector{N}
-              ) where {N}
-    return A[e1.i, e2.i] * e1.v * e2.v
-end
-
-"""
     BoxDirections{N} <: AbstractDirections{N}
 
 Box direction representation.
@@ -100,7 +57,7 @@ function Base.iterate(bd::BoxDirections{N}, state::Int=1) where {N}
     if state == 0
         return nothing
     end
-    vec = UnitVector{N}(abs(state), bd.n, convert(N, sign(state)))
+    vec = SingleEntryVector(abs(state), bd.n, convert(N, sign(state)))
     state = (state == bd.n) ? -bd.n : state + 1
     return (vec, state)
 end
