+++ src/helper_functions.jl
@@ -1,36 +1,3 @@
-export _At_mul_B
-
-# default tolerance for matrix condition number (see 'isinvertible')
-const DEFAULT_COND_TOL = 1e6
-
-# matrix-matrix multiplication and division
-@inline _At_mul_B(A, B) = transpose(A) * B
-@inline _At_ldiv_B(A, B) = transpose(A) \ B
-
-"""
-    dot_zero(x::AbstractVector{N}, y::AbstractVector{N}) where{N<:Real}
-
-Dot product with preference for zero value in the presence of infinity values.
-
-### Input
-
-- `x` -- first vector
-- `y` -- second vector
-
-### Output
-
-The dot product of `x` and `y`, but with the rule that `0 * Inf == 0`.
-"""
-function dot_zero(x::AbstractVector{N}, y::AbstractVector{N}) where{N<:Real}
-    res = zero(N)
-    for i in 1:length(x)
-        if !iszero(x[i]) && !iszero(y[i])
-            res += x[i] * y[i]
-        end
-    end
-    return res
-end
-
 """
     sign_cadlag(x::N)::N where {N<:Real}
 
@@ -62,231 +29,6 @@ julia> LazySets.sign_cadlag.([-0.6, 1.3, 0.0])
 """
 function sign_cadlag(x::N)::N where {N<:Real}
     return x < zero(x) ? -one(x) : one(x)
-end
-
-"""
-    ispermutation(u::AbstractVector{T}, v::AbstractVector)::Bool where {T}
-
-Check that two vectors contain the same elements up to reordering.
-
-### Input
-
-- `u` -- first vector
-- `v` -- second vector
-
-### Output
-
-`true` iff the vectors are identical up to reordering.
-
-### Examples
-
-```jldoctest
-julia> LazySets.ispermutation([1, 2, 2], [2, 2, 1])
-true
-
-julia> LazySets.ispermutation([1, 2, 2], [1, 1, 2])
-false
-
-```
-"""
-function ispermutation(u::AbstractVector{T}, v::AbstractVector)::Bool where {T}
-    if length(u) != length(v)
-        return false
-    end
-    occurrence_map = Dict{T, Int}()
-    has_duplicates = false
-    for e in u
-        if e ∉ v
-            return false
-        end
-        if haskey(occurrence_map, e)
-            occurrence_map[e] += 1
-            has_duplicates = true
-        else
-            occurrence_map[e] = 1
-        end
-    end
-    if has_duplicates
-        for e in v
-            if !haskey(occurrence_map, e) || occurrence_map[e] == 0
-                return false
-            end
-            occurrence_map[e] -= 1
-        end
-    end
-    return true
-end
-
-"""
-    issquare(M::AbstractMatrix)::Bool
-
-Check whether a matrix is square.
-
-### Input
-
-- `M` -- matrix
-
-### Output
-
-`true` iff the matrix is square.
-"""
-function issquare(M::AbstractMatrix)::Bool
-    m, n = size(M)
-    return m == n
-end
-
-"""
-    isinvertible(M::Matrix; [cond_tol]::Number=DEFAULT_COND_TOL)
-
-A sufficient check of a matrix being invertible (or nonsingular).
-
-### Input
-
-- `M`        -- matrix
-- `cond_tol` -- (optional, default: `DEFAULT_COND_TOL`) tolerance of matrix
-                condition
-
-### Output
-
-If the result is `true`, `M` is invertible.
-If the result is `false`, the matrix is non-square or this function could not
-conclude.
-
-### Algorithm
-
-We check whether the matrix is square and whether the
-[matrix condition number](https://en.wikipedia.org/wiki/Condition_number#Matrices)
-`cond(M)` is below some prescribed tolerance.
-"""
-function isinvertible(M::Matrix; cond_tol::Number=DEFAULT_COND_TOL)
-    return issquare(M) && cond(M) < cond_tol
-end
-
-# cond is not available for sparse matrices; see JuliaLang#6485 and related issues
-function isinvertible(M::AbstractSparseMatrix;
-                      cond_tol::Number=DEFAULT_COND_TOL)
-    return issquare(M) && isinvertible(Matrix(M), cond_tol=cond_tol)
-end
-
-function isinvertible(M::Diagonal; cond_tol=nothing)
-    return !any(iszero, diag(M))
-end
-
-"""
-    remove_duplicates_sorted!(v::AbstractVector)
-
-Remove duplicate entries in a sorted vector.
-
-### Input
-
-- `v` -- sorted vector
-
-### Output
-
-The input vector without duplicates.
-"""
-function remove_duplicates_sorted!(v::AbstractVector)
-    for i in length(v)-1:-1:1
-        if v[i] == v[i+1]
-            splice!(v, i+1)
-        end
-    end
-    return v
-end
-
-"""
-    samedir(u::AbstractVector{N},
-            v::AbstractVector{N})::Tuple{Bool, Real} where {N<:Real}
-
-Check whether two vectors point in the same direction.
-
-### Input
-
-- `u` -- first vector
-- `v` -- second vector
-
-### Output
-
-`(true, k)` iff the vectors are identical up to a positive scaling factor `k`,
-and `(false, 0)` otherwise.
-
-
-### Examples
-
-```jldoctest
-julia> using LazySets: samedir
-
-julia> samedir([1, 2, 3], [2, 4, 6])
-(true, 0.5)
-
-julia> samedir([1, 2, 3], [3, 2, 1])
-(false, 0)
-
-julia> samedir([1, 2, 3], [-1, -2, -3])
-(false, 0)
-
-```
-"""
-function samedir(u::AbstractVector{N},
-                 v::AbstractVector{N}
-                )::Tuple{Bool, Real} where {N<:Real}
-    @assert length(u) == length(v) "wrong dimension"
-    no_factor = true
-    factor = 0
-    @inbounds for i in 1:length(u)
-        if u[i] == 0
-            if v[i] != 0
-                return (false, 0)
-            end
-            continue
-        elseif v[i] == 0
-            return (false, 0)
-        end
-        if no_factor
-            no_factor = false
-            factor = u[i] / v[i]
-            if factor < 0
-                return (false, 0)
-            end
-        elseif factor != u[i] / v[i]
-            return (false, 0)
-        end
-    end
-    if no_factor
-        # both vectors are zero
-        return (true, 0)
-    end
-    return (true, factor)
-end
-
-"""
-    nonzero_indices(v::AbstractVector{N})::Vector{Int} where {N<:Real}
-
-Return the indices in which a vector is non-zero.
-
-### Input
-
-- `v` -- vector
-
-### Output
-
-A vector of ascending indices `i` such that the vector is non-zero in dimension
-`i`.
-"""
-function nonzero_indices(v::AbstractVector{N})::Vector{Int} where {N<:Real}
-    n = length(v)
-    res = Vector{Int}()
-    sizehint!(res, n)
-    for i in 1:n
-        if v[i] != zero(N)
-            push!(res, i)
-        end
-    end
-    return res
-end
-
-function nonzero_indices(v::SparseVector{N})::Vector{Int} where {N<:Real}
-    return v.nzind
 end
 
 """
@@ -358,50 +100,6 @@ function reseed(rng::AbstractRNG, seed::Union{Int, Nothing})::AbstractRNG
         return Random.seed!(rng, seed)
     end
     return rng
-end
-
-"""
-    cross_product(M::AbstractMatrix{N}) where {N<:Real}
-
-Compute the high-dimensional cross product of ``n-1`` ``n``-dimensional vectors.
-
-### Input
-
-- `M` -- ``n × n - 1``-dimensional matrix
-
-### Output
-
-A vector.
-
-### Algorithm
-
-The cross product is defined as follows:
-
-```math
-\\left[ \\dots, (-1)^{n+1} \\det(M^{[i]}), \\dots \\right]^T
-```
-where ``M^{[i]}`` is defined as ``M`` with the ``i``-th row removed.
-See *Althoff, Stursberg, Buss: Computing Reachable Sets of Hybrid Systems Using
-a Combination of Zonotopes and Polytopes. 2009.*
-"""
-function cross_product(M::AbstractMatrix{N})::Vector{N} where {N<:Real}
-    n = size(M, 1)
-    @assert 1 < n == size(M, 2) + 1 "the matrix must be n x (n-1) dimensional"
-
-    v = Vector{N}(undef, n)
-    minus = false
-    for i in 1:n
-        Mi = view(M, 1:n .!= i, :)  # remove i-th row
-        d = det(Mi)
-        if minus
-            v[i] = -d
-            minus = false
-        else
-            v[i] = d
-            minus = true
-        end
-    end
-    return v
 end
 
 """
@@ -563,105 +261,4 @@ function subtypes(interface, concrete::Bool)
         end
     end
     return sort(result, by=string)
-end
-
-"""
-    delete_zero_columns(A::AbstractMatrix)
-
-Remove all columns that only contain zeros from a given matrix.
-
-### Input
-
-- `A`    -- matrix
-- `copy` -- (optional, default: `false`) flag to copy the matrix
-
-### Output
-
-A matrix.
-
-If the input matrix `A` does not contain any zero column, we return `A` unless
-the option `copy` is set.
-If the input matrix contains zero columns, we always return a copy if the option
-`copy` is set and otherwise a `SubArray` via `@view`.
-"""
-function delete_zero_columns(A::AbstractMatrix, copy::Bool=false)
-    n = size(A, 2)
-    nonzero_columns = Vector{Int}()
-    sizehint!(nonzero_columns, n)
-    for i in 1:n
-        if !iszero(A[:, i])
-            push!(nonzero_columns, i)
-        end
-    end
-    if copy
-        if length(nonzero_columns) == n
-            return copy(A)
-        else
-            return A[:, nonzero_columns]
-        end
-    else
-        if length(nonzero_columns) == n
-            return A
-        else
-            return @view A[:, nonzero_columns]
-        end
-    end
-end
-
-"""
-    inner(x::AbstractVector{N}, A::AbstractMatrix{N}, y::AbstractVector{N}
-         ) where {N}
-
-Compute the inner product ``xᵀ A y``.
-
-### Input
-
-- `x` -- vector on the left
-- `A` -- matrix
-- `y` -- vector on the right
-
-### Output
-
-The (scalar) result of the multiplication.
-"""
-function inner(x::AbstractVector{N}, A::AbstractMatrix{N}, y::AbstractVector{N}
-              ) where {N}
-    dot(x, A * y)
-end
-
-"""
-    rectify(x::AbstractVector{N}) where {N<:Real}
-
-Rectify a vector, i.e., take the element-wise maximum with zero.
-
-### Input
-
-- `x` -- vector
-
-### Output
-
-A copy of the vector where each negative entry is replaced by zero.
-"""
-function rectify(x::AbstractVector{N}) where {N<:Real}
-    return map(xi -> max(xi, zero(N)), x)
-end
-
-"""
-    iscounterclockwise(result, correct_expr) where {N<:Real}
-
-Returns a boolean, true if the elements in `result` are ordered in
-a couter-clockwise fashion and in the same order as `correct_expr`.
-
-### Input
-
-- `result` -- vector
-- `correct_expr` -- paragon vector with elements in ccw order
-
-### Output
-
-A boolean indicating if the elements of `result` are in the same order as
-`correct_expr` or any of its cyclic permutations.
-"""
-function iscounterclockwise(result, correct_expr)
-    return any([result == circshift(correct_expr, i) for i in 0:length(result)-1])
 end
